+++ DWH/IFS/Tables/PURCHASE_PART_SUPPLIER.sql
@@ -12,7 +12,7 @@
     [CUM_RCPTS_QTY]               INT             NULL,
     [DISCOUNT]                    INT             NULL,
     [INTERNAL_CONTROL_TIME]       INT             NULL,
-    [LIST_PRICE]                  INT             NULL,
+    [LIST_PRICE]                  DECIMAL (18, 2) NULL,
     [MINIMUM_QTY]                 INT             NULL,
     [PRICE_CONV_FACTOR]           INT             NULL,
     [PRICE_UNIT_MEAS]             NVARCHAR (10)   NULL,
@@ -66,6 +66,8 @@
     [ProcessLogID]                BIGINT          NULL,
     CONSTRAINT [PK_PURCHASE_PART_SUPPLIER] PRIMARY KEY CLUSTERED ([CONTRACT] ASC, [Eff_Date] ASC, [PART_NO] ASC, [VENDOR_NO] ASC) ON [DWH_IFS]
 ) ON [DWH_IFS];
+
+
