+++ DWH/IFS/Tables/SALES_PRICE_LIST_PART.sql
@@ -4,10 +4,10 @@
     [MIN_QUANTITY]      INT             NOT NULL,
     [VALID_FROM_DATE]   DATETIME2 (7)   NOT NULL,
     [BASE_PRICE_SITE]   NVARCHAR (5)    NULL,
-    [BASE_PRICE]        INT             NULL,
+    [BASE_PRICE]        DECIMAL (18, 2) NULL,
     [PERCENTAGE_OFFSET] INT             NULL,
     [AMOUNT_OFFSET]     INT             NULL,
-    [SALES_PRICE]       INT             NULL,
+    [SALES_PRICE]       DECIMAL (18, 2) NULL,
     [ROUNDING]          INT             NULL,
     [LAST_UPDATED]      DATETIME2 (7)   NULL,
     [DISCOUNT]          INT             NULL,
@@ -22,6 +22,8 @@
     [ProcessLogID]      BIGINT          NULL,
     CONSTRAINT [PK_SALES_PRICE_LIST_PART] PRIMARY KEY CLUSTERED ([CATALOG_NO] ASC, [Eff_Date] ASC, [MIN_QUANTITY] ASC, [PRICE_LIST_NO] ASC, [VALID_FROM_DATE] ASC) ON [DWH_IFS]
 ) ON [DWH_IFS];
+
+
