+++ Staging/IFS/Tables/EXT_DESADV_LINE.sql
@@ -6,7 +6,7 @@
     [LINE_NO]                  NVARCHAR (2000) NULL,
     [RELEASE_NO]               NVARCHAR (2000) NULL,
     [PART_NO]                  NVARCHAR (2000) NULL,
-    [DESPATCH_QTY]             INT             NULL,
+    [DESPATCH_QTY]             BIGINT          NULL,
     [SALES_UNIT_MEAS]          NVARCHAR (2000) NULL,
     [SALES_PART_NO]            NVARCHAR (2000) NULL,
     [SUPPLIER_ORDER_NO]        NVARCHAR (2000) NULL,
@@ -42,4 +42,6 @@
     [BatchID]                  BIGINT          NULL,
     [ProcessLogID]             BIGINT          NULL
 );
+
+
