+++ Staging/IFS/Tables/PURCHASE_PART_SUPPLIER.sql
@@ -12,7 +12,7 @@
     [CUM_RCPTS_QTY]               INT             NULL,
     [DISCOUNT]                    INT             NULL,
     [INTERNAL_CONTROL_TIME]       INT             NULL,
-    [LIST_PRICE]                  INT             NULL,
+    [LIST_PRICE]                  DECIMAL (18, 2) NULL,
     [MINIMUM_QTY]                 INT             NULL,
     [PRICE_CONV_FACTOR]           INT             NULL,
     [PRICE_UNIT_MEAS]             NVARCHAR (10)   NULL,
@@ -61,4 +61,6 @@
     [BatchID]                     BIGINT          NULL,
     [ProcessLogID]                BIGINT          NULL
 );
+
+
