+++ Staging/IFS/Tables/ROUTING_OPERATION.sql
@@ -13,7 +13,7 @@
     [LOT_QTY]                   INT             NULL,
     [EFFICIENCY_FACTOR]         INT             NULL,
     [MACH_RUN_FACTOR]           INT             NULL,
-    [MACH_SETUP_TIME]           INT             NULL,
+    [MACH_SETUP_TIME]           DECIMAL (18, 2) NULL,
     [MOVE_TIME]                 INT             NULL,
     [LABOR_RUN_FACTOR]          INT             NULL,
     [LABOR_SETUP_TIME]          INT             NULL,
@@ -46,4 +46,6 @@
     [BatchID]                   BIGINT          NULL,
     [ProcessLogID]              BIGINT          NULL
 );
+
+
