+++ Staging/IFS/Tables/SALES_PRICE_LIST_PART.sql
@@ -4,10 +4,10 @@
     [MIN_QUANTITY]      INT             NOT NULL,
     [VALID_FROM_DATE]   DATETIME2 (7)   NOT NULL,
     [BASE_PRICE_SITE]   NVARCHAR (5)    NULL,
-    [BASE_PRICE]        INT             NULL,
+    [BASE_PRICE]        DECIMAL (18, 2) NULL,
     [PERCENTAGE_OFFSET] INT             NULL,
     [AMOUNT_OFFSET]     INT             NULL,
-    [SALES_PRICE]       INT             NULL,
+    [SALES_PRICE]       DECIMAL (18, 2) NULL,
     [ROUNDING]          INT             NULL,
     [LAST_UPDATED]      DATETIME2 (7)   NULL,
     [DISCOUNT]          INT             NULL,
@@ -17,4 +17,6 @@
     [BatchID]           BIGINT          NULL,
     [ProcessLogID]      BIGINT          NULL
 );
+
+
