+++ Staging/IFS/Tables/SHIPMENT_PACKAGE_UNIT.sql
@@ -15,8 +15,8 @@
     [REL_NO]                 NVARCHAR (4)    NULL,
     [LINE_ITEM_NO]           INT             NULL,
     [SERIAL_NO]              NVARCHAR (50)   NULL,
-    [NET_WEIGHT]             INT             NULL,
-    [HANDLING_UNIT_WEIGHT]   INT             NULL,
+    [NET_WEIGHT]             DECIMAL (18, 2) NULL,
+    [HANDLING_UNIT_WEIGHT]   DECIMAL (18, 2) NULL,
     [VOLUME]                 INT             NULL,
     [TYPE_OF_GOODS]          NVARCHAR (35)   NULL,
     [SSCC]                   NVARCHAR (18)   NULL,
@@ -32,4 +32,6 @@
     [BatchID]                BIGINT          NULL,
     [ProcessLogID]           BIGINT          NULL
 );
+
+
