+++ Staging/IFS/Tables/SHOP_ORDER_OPERATION.sql
@@ -6,7 +6,7 @@
     [OPERATION_DESCRIPTION]     NVARCHAR (35)   NULL,
     [EFFICIENCY_FACTOR]         INT             NULL,
     [MACH_RUN_FACTOR]           INT             NULL,
-    [MACH_SETUP_TIME]           INT             NULL,
+    [MACH_SETUP_TIME]           DECIMAL (18, 2) NULL,
     [MOVE_TIME]                 INT             NULL,
     [QUEUE_TIME]                INT             NULL,
     [CBS_QUEUE_TIME]            INT             NULL,
@@ -76,6 +76,8 @@
     [BatchID]                   BIGINT          NULL,
     [ProcessLogID]              BIGINT          NULL
 );
+
+
