+++ StagingIFS/PROD/Tables/INVENTORY_TRANSACTION_HIST.sql
@@ -17,7 +17,7 @@
     [REJECT_CODE]                    NVARCHAR (8)    NULL,
     [TRANSACTION]                    NVARCHAR (4000) NULL,
     [PRE_ACCOUNTING_ID]              INT             NULL,
-    [COST]                           INT             NULL,
+    [COST]                           DECIMAL (18, 2) NULL,
     [DATE_APPLIED]                   DATETIME2 (7)   NULL,
     [DIRECTION]                      NVARCHAR (1)    NULL,
     [ORDER_TYPE]                     NVARCHAR (4000) NULL,
@@ -76,4 +76,6 @@
     [ProcessLogID]                   BIGINT          NULL,
     CONSTRAINT [PK_INVENTORY_TRANSACTION_HIST] PRIMARY KEY CLUSTERED ([TRANSACTION_ID] ASC)
 );
+
+
